Split PrecompiledScriptPluginAccessorsTest
