+++ subprojects/kotlin-dsl-plugins/src/integTest/kotlin/org/gradle/kotlin/dsl/plugins/precompiled/PrecompiledScriptPluginAccessorsTest.kt
@@ -62,97 +62,6 @@ import java.io.File
 
 @LeaksFileHandles("Kotlin Compiler Daemon working directory")
 class PrecompiledScriptPluginAccessorsTest : AbstractPrecompiledScriptPluginTest() {
-
-    @Test
-    @ToBeFixedForConfigurationCache
-    fun `settings and init scripts are not evaluated when generating accessors`() {
-        // given:
-        val evaluationLog = file("evaluation.log")
-        withFolders {
-            // a precompiled script plugin contributing an extension
-            "producer/src/main/kotlin" {
-                withFile(
-                    "producer.plugin.gradle.kts",
-                    """extensions.add("answer", 42)"""
-                )
-            }
-            // a consumer of the precompiled script plugin extension
-            "consumer/src/main/kotlin" {
-                withFile(
-                    "consumer.plugin.gradle.kts",
-                    """
-                        plugins { id("producer.plugin") }
-                        println(answer)
-                    """
-                )
-            }
-        }
-        withDefaultSettings().appendText(
-            """
-                include("producer", "consumer")
-                file("${evaluationLog.normalisedPath}").appendText("<settings>")
-            """
-        )
-        withKotlinDslPlugin().appendText(
-            """
-                subprojects {
-                    apply(plugin = "org.gradle.kotlin.kotlin-dsl")
-                    $repositoriesBlock
-                }
-                project(":consumer") {
-                    dependencies {
-                        implementation(project(":producer"))
-                    }
-                }
-            """
-        )
-
-        // and: a bunch of init scripts
-        fun initScript(file: File, label: String) = file.apply {
-            parentFile.mkdirs()
-            writeText("file('${evaluationLog.normalisedPath}') << '$label'")
-        }
-
-        val gradleUserHome = newDir("gradle-user-home")
-
-        // <user-home>/init.gradle
-        initScript(
-            gradleUserHome.resolve("init.gradle"),
-            "<init>"
-        )
-        // <user-home>/init.d/init.gradle
-        initScript(
-            gradleUserHome.resolve("init.d/init.gradle"),
-            "<init.d>"
-        )
-        // -I init.gradle
-        val initScript = initScript(
-            file("init.gradle"),
-            "<command-line>"
-        )
-
-        // when: precompiled script plugin accessors are generated
-        buildWithGradleUserHome(
-            gradleUserHome,
-            "generatePrecompiledScriptPluginAccessors",
-            "-I",
-            initScript.absolutePath
-        ).apply {
-            // then: the settings and init scripts are only evaluated once by the outer build
-            assertThat(
-                evaluationLog.text,
-                equalTo("<command-line><init><init.d><settings>")
-            )
-        }
-    }
-
-    private
-    fun buildWithGradleUserHome(gradleUserHomeDir: File, vararg arguments: String) =
-        gradleExecuterFor(arguments)
-            .withGradleUserHomeDir(gradleUserHomeDir)
-            .withOwnUserHomeServices()
-            .run()
-
     @Test
     @ToBeFixedForConfigurationCache
     fun `cannot use type-safe accessors for extensions contributed in afterEvaluate`() {
@@ -201,20 +110,6 @@ class PrecompiledScriptPluginAccessorsTest : AbstractPrecompiledScriptPluginTest
         buildAndFail("compileKotlin").apply {
             assertHasCause("Compilation error.")
             assertHasErrorOutput("Unresolved reference: after")
-        }
-    }
-
-    @Test
-    @ToBeFixedForConfigurationCache
-    fun `generated type-safe accessors suppress deprecation warnings`() {
-        // `java-gradle-plugin` adds deprecated task `ValidateTaskProperties`
-        givenPrecompiledKotlinScript(
-            "java-project.gradle.kts",
-            """
-            plugins { `java-gradle-plugin` }
-            """
-        ).apply {
-            assertNotOutput("'ValidateTaskProperties' is deprecated.")
         }
     }
